Increase default values
Up max concurrent sends to 3, this shouldn't really cause
any problems still. It should help in some cases
where the server is under pressure and it simply needs to queue
up chunks a bit to meet the send rate required.

Concurrent loads is an interesting case, looks like it's best
this value is per player so that each player can get chunks
loading. I noticed for a some people using the defaults,
the chunk system simply wasn't being fully utilised. Now
that should be fixed, and it should be fully utilised now.